requirements.txt
anyio==3.3.4
asgiref==3.4.1
bert-extractive-summarizer==0.8.1
blis==0.7.5
boto3==1.19.2
botocore==1.22.2
catalogue==2.0.6
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
cymem==2.0.6
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.1.0/en_core_web_md-3.1.0-py3-none-any.whl
fastapi==0.70.0
filelock==3.3.1
h11==0.12.0
huggingface-hub==0.0.19
idna==3.3
Jinja2==3.0.2
jmespath==0.10.0
joblib==1.1.0
MarkupSafe==2.0.1
murmurhash==1.0.6
neuralcoref==4.0
numpy==1.21.3
packaging==21.0
pathy==0.6.0
Pillow==8.4.0
preshed==3.0.6
pydantic==1.8.2
pyparsing==3.0.1
python-dateutil==2.8.2
PyYAML==6.0
regex==2021.10.23
requests==2.26.0
s3transfer==0.5.0
sacremoses==0.0.46
scikit-learn==1.0.1
scipy==1.7.1
six==1.16.0
smart-open==5.2.1
sniffio==1.2.0
spacy==3.1.3
spacy-legacy==3.0.8
srsly==2.4.2
starlette==0.16.0
thinc==8.0.11
threadpoolctl==3.0.0
tokenizers==0.10.3
torch
tqdm==4.62.3
transformers==4.11.3
typer==0.4.0
typing-extensions==3.10.0.2
urllib3==1.26.7
uvicorn==0.15.0
wasabi==0.8.2
